fix: full optimize front resource
