+++ app/src/main/resources/static/assets/css/community/editorjs.css
