+++ app/src/main/resources/static/assets/css/dictionary/autocomplete-theme-bootstrap.css
