--- app/src/main/resources/static/assets/js/cmmn/aos.mjs
@@ -1,3 +1,0 @@
-AOS.init(
-    {duration: 250, offset:20, delay: 0, once: true}
-);
