+++ app/src/main/resources/static/assets/js/cmmn/toast.mjs
@@ -1,10 +1,10 @@
 document.addEventListener("htmx:afterOnLoad", function () {
-    var toastElList = [].slice.call(document.querySelectorAll('.toast'))
-    var toastList = toastElList.map(function (toastEl) {
+    let toastElList = [].slice.call(document.querySelectorAll('.toast'))
+    let toastList = toastElList.map(function (toastEl) {
         return new bootstrap.Toast(toastEl, { autohide: true })
     });
 
-    var toastMessage = document.getElementById('toast-message');
+    let toastMessage = document.getElementById('toast-message');
     if (toastMessage && toastMessage.textContent.trim() !== '') {
         toastList.forEach(toast => toast.show());
     }
