+++ app/src/main/resources/static/assets/js/dictionary/aos.mjs
@@ -1,0 +1,5 @@
+import AOS from 'https://cdn.jsdelivr.net/npm/aos/+esm';
+
+export function init() {
+    AOS.init({duration: 250, offset: 20, delay: 0, once: true});
+}
