+++ app/src/main/resources/templates/base/layout.html
@@ -3,7 +3,7 @@
 
 <head>
     <meta charset="utf-8">
-<link rel="manifest" href="/assets/js/manifest.json" crossorigin="use-credentials">
+<link rel="manifest" href="/assets/js/manifest.json">
     <meta name="viewport" content="width=device-width, shrink-to-fit=0, user-scalable=0, initial-scale=1">
     <title>분리ㅅㄱ</title>
     <meta name="theme-color" content="#ffffff">
@@ -24,30 +24,10 @@
     <link rel="icon" type="image/png" sizes="192x192" href="/assets/img/favicon/192x192.png">
     <link rel="icon" type="image/png" sizes="512x512" href="/assets/img/favicon/512x512.png">
     <link rel="stylesheet" href="/assets/bootstrap/css/bootstrap.min.css">
-    <link rel="stylesheet" href="/assets/css/cmmn/autocomplete-theme-bootstrap.css">
-    <link rel="stylesheet" href="/assets/css/cmmn/editor-js.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/aos/dist/aos.css">
+    <link rel="stylesheet" href="/assets/css/dictionary/aos.css">
     <link rel="stylesheet" href="/assets/css/cmmn/style.css">
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/editorjs@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/list@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/header@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/table@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/image@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/paragraph@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/embed@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/code@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/quote@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/delimiter@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/simple-image@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/warning@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/link@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/attaches@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/checklist@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/inline-code@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/marker@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/raw@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@editorjs/image@latest"></script>
-<script src="https://cdn.jsdelivr.net/npm/@weekwood/editorjs-video@latest"></script>
+    <link rel="stylesheet" href="/assets/css/community/editorjs.css">
+    <link rel="stylesheet" href="/assets/css/dictionary/autocomplete-theme-bootstrap.css">
 </head>
 
 <body oncontextmenu="return false" ondragstart="return false">
@@ -57,17 +37,10 @@
     <script src="/assets/js/jquery.min.js"></script>
     <script src="/assets/bootstrap/js/bootstrap.min.js"></script>
     <script src="/assets/js/bs-init.js"></script>
-    <script src="/assets/js/cmmn/algolia.mjs" type="module"></script>
-    <script src="/assets/js/cmmn/aos.mjs" type="module"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@algolia/autocomplete-js" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/algoliasearch" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/aos/dist/aos.js" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/htmx.org" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/timeago/jquery.timeago.min.js"></script>
+    <script src="/assets/js/app.mjs" type="module"></script>
+    <script src="https://cdn.jsdelivr.net/npm/htmx.org"></script>
     <script src="/assets/js/cmmn/theme-switch.mjs" type="module"></script>
-    <script src="/assets/js/cmmn/time-ago.js"></script>
     <script src="/assets/js/cmmn/toast.mjs" type="module"></script>
-    <script src="/assets/js/community/post-color-changer.mjs" type="module"></script>
 </body>
 
 </html>
