+++ app/src/main/resources/templates/community/board.html
@@ -24,7 +24,6 @@
     <link rel="icon" type="image/png" sizes="512x512" href="/assets/img/favicon/512x512.png">
     <link rel="stylesheet" href="/assets/bootstrap/css/bootstrap.min.css">
     <link rel="stylesheet" href="/assets/css/animate.min.css">
-    <link rel="stylesheet" href="/assets/css/cmmn/style.css">
 </head>
 
 <body>
