+++ app/src/main/resources/templates/dictionary/info.html
@@ -23,7 +23,6 @@
     <link rel="icon" type="image/png" sizes="192x192" href="/assets/img/favicon/192x192.png">
     <link rel="icon" type="image/png" sizes="512x512" href="/assets/img/favicon/512x512.png">
     <link rel="stylesheet" href="/assets/bootstrap/css/bootstrap.min.css">
-    <link rel="stylesheet" href="/assets/css/cmmn/style.css">
 </head>
 
 <body>
